refactor(ui): 과도한 추상화 로직 제거
