+++ packages/ui/src/components/mdx/index.ts
@@ -1,9 +1,8 @@
 export { MDX } from "./MDX";
 export { useSerializedMDX } from "./hooks/use-serialized-mdx";
-export { serialize, createDefaultSerializeConfig } from "./lib/serializer";
+export { serialize, createDefaultSerializeConfig } from "./lib/serializer/serialize";
 export { createMDXComponentMap, useMDXComponentMap } from "./lib/factory/create-component-map";
-export { createRuntimeConfig, createSerializeConfig, mergeSerializeConfig } from "./lib/builders";
-export { isExternalUrl, withRuntimeConfig, mergeComponents } from "./lib/utils";
+export { isExternalUrl, withRuntimeConfig } from "./lib/utils";
 export type {
   MDXRuntimeConfig,
   SerializeConfig,
