--- packages/ui/src/components/mdx/lib/builders/create-runtime-config.ts
@@ -1,16 +1,0 @@
-import type { ComponentType, ReactNode } from "react";
-import type { MDXRuntimeConfig } from "../../types";
-
-/**
- * 순수 함수: MDXRuntimeConfig 생성
- */
-export function createRuntimeConfig(
-  config: MDXRuntimeConfig
-): MDXRuntimeConfig {
-  return {
-    LinkComponent: config.LinkComponent,
-    processImageSource: config.processImageSource,
-    appName: config.appName,
-  };
-}
-
