--- packages/ui/src/components/mdx/lib/builders/index.ts
@@ -1,3 +1,0 @@
-export { createRuntimeConfig } from "./create-runtime-config";
-export { createSerializeConfig, mergeSerializeConfig } from "./create-serialize-config";
-
