--- packages/ui/src/components/mdx/lib/serializer/index.ts
@@ -1,4 +1,0 @@
-export { serialize } from "./serialize";
-export { createDefaultSerializeConfig } from "./default-config";
-export type { SerializeConfig } from "../../types";
-
