+++ packages/ui/src/components/mdx/lib/serializer/serialize.ts
@@ -1,5 +1,30 @@
 import { serialize as serializeMDX } from "next-mdx-remote/serialize";
+import rehypeSlug from "rehype-slug";
+import rehypeAutolinkHeadings from "rehype-autolink-headings";
+import rehypePrettyCode from "rehype-pretty-code";
 import type { SerializeConfig } from "../../types";
+
+/**
+ * 순수 함수: 기본 serialize 설정 생성
+ */
+export function createDefaultSerializeConfig(): SerializeConfig {
+  return {
+    rehypePlugins: [
+      rehypeSlug,
+      [
+        rehypeAutolinkHeadings,
+        {
+          behavior: "wrap",
+          properties: {
+            className: ["anchor"],
+            ariaLabel: "anchor",
+          },
+        },
+      ],
+      [rehypePrettyCode, { theme: "dark-plus" }],
+    ],
+  };
+}
 
 /**
  * 순수 함수: MDX 소스를 serialize
@@ -17,4 +42,3 @@ export async function serialize(
     },
   });
 }
-
