+++ packages/ui/src/components/mdx/lib/utils/index.ts
@@ -1,4 +1,2 @@
-export { isExternalUrl } from "./is-external-url";
-export { withRuntimeConfig } from "./with-runtime-config";
-export { mergeComponents } from "./merge-components";
+export { isExternalUrl, withRuntimeConfig } from "./utils";
